+++ packages/node/src/indexer/blockDispatcher/worker-block-dispatcher.service.ts
@@ -24,6 +24,7 @@ import {
   dynamicDsHostFunctions,
   IProjectUpgradeService,
   HostUnfinalizedBlocks,
+  PoiSyncService,
 } from '@subql/node-core';
 import { Store } from '@subql/types-core';
 import {
@@ -87,6 +88,7 @@ export class WorkerBlockDispatcherService
     storeService: StoreService,
     storeCacheService: StoreCacheService,
     poiService: PoiService,
+    poiSyncService: PoiSyncService,
     @Inject('ISubqueryProject') project: SubqueryProject,
     dynamicDsService: DynamicDsService,
     unfinalizedBlocksSevice: UnfinalizedBlocksService,
@@ -101,6 +103,7 @@ export class WorkerBlockDispatcherService
       storeService,
       storeCacheService,
       poiService,
+      poiSyncService,
       project,
       dynamicDsService,
       () =>