+++ packages/node/src/indexer/project.service.test.ts
@@ -33,6 +33,7 @@ describe('ProjectService', () => {
       null,
       null,
       null,
+      null,
       {} as any,
       null,
       null,