+++ packages/node/src/indexer/project.service.ts
@@ -6,6 +6,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   PoiService,
+  PoiSyncService,
   BaseProjectService,
   StoreService,
   NodeConfig,
@@ -36,6 +37,8 @@ export class ProjectService extends BaseProjectService<
     dsProcessorService: DsProcessorService,
     @Inject(ApiService) apiService: EthereumApiService,
     @Inject(isMainThread ? PoiService : 'Null') poiService: PoiService,
+    @Inject(isMainThread ? PoiSyncService : 'Null')
+    poiSyncService: PoiSyncService,
     @Inject(isMainThread ? Sequelize : 'Null') sequelize: Sequelize,
     @Inject('ISubqueryProject') project: SubqueryProject,
     @Inject('IProjectUpgradeService')
@@ -50,6 +53,7 @@ export class ProjectService extends BaseProjectService<
       dsProcessorService,
       apiService,
       poiService,
+      poiSyncService,
       sequelize,
       project,
       projectUpgradeService,