+++ packages/node/src/init.ts
@@ -44,7 +44,7 @@ export async function bootstrap(): Promise<void> {
 
   try {
     const app = await NestFactory.create(AppModule, {
-      logger: new NestLogger(debug),
+      logger: new NestLogger(!!debug),
     });
     await app.init();