+++ packages/node/src/subcommands/testing.service.ts
@@ -49,7 +49,7 @@ export class TestingService extends BaseTestingService<
     const testContext = await NestFactory.createApplicationContext(
       TestingModule,
       {
-        logger: new NestLogger(this.nodeConfig.debug),
+        logger: new NestLogger(!!this.nodeConfig.debug),
       },
     );