+++ packages/node/CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- debug has changed from a boolean to a string to allow scoping debug log level (#2077)
+
+### Fixed
+- Sync with node-core.
+  - Fixed Poi migration performance issue.
+  - Fixed AutoQueue timeout issue.
+  - Fixed Poi sync could block DB IO and drop connection issue.
+
 ## [3.0.3] - 2023-10-05
 ### Changed
 - Version bump with `@subql/types-ethereum` 3.0.2